+++ src/database/jsondb/types.ts
@@ -6,12 +6,14 @@ export interface AliasesType {
     ccCommands: Record<string, string[]>;
     ccControllers: Record<string, number>;
     commands: Record<string, Command>;
+    macros: Record<string, Array<[command: string, timeout: number]>>;
 }
 
-export const CHORD_PROGRESSIONS = 'chordProgressions';
-export const CC_COMMANDS = 'ccCommands';
-export const CC_CONTROLLERS = 'ccControllers';
+export const CHORD_PROGRESSIONS_KEY = 'chordProgressions';
+export const CC_COMMANDS_KEY = 'ccCommands';
+export const CC_CONTROLLERS_KEY = 'ccControllers';
 export const COMMANDS_KEY = 'commands';
+export const MACROS_KEY = 'macros';
 
 export interface RewardsType {
     rewardTitleCommandMap: Record<string, [command: string, cost: number]>;
@@ -23,10 +25,10 @@ export interface PermissionsType {
     permissionsMap: Record<Command, PermissionsTable>;
 }
 
+export const PERMISSIONS_MAP = 'permissionsMap';
+
 export interface PermissionsTable {
     requirements: Array<keyof UserRoles>;
     whitelist: string[];
     blacklist: string[];
 }
-
-export const PERMISSIONS_MAP = 'permissionsMap';