+++ src/twitch/chat/handler.ts
@@ -1,11 +1,12 @@
 import { ChatClient } from '@twurple/chat';
-import { CONFIG, ERROR_MSG, GLOBAL } from '../../configuration/constants';
+import { CONFIG, ERROR_MSG } from '../../configuration/constants';
 import { CommandHandlerType, MessageHandler, RequestSource, TwitchParams } from './types';
-import { getCommand } from '../../command/utils';
+import { getCommandList } from '../../command/utils';
 import * as CommandHandlers from '../../command/handler';
 import { checkCommandAccess } from '../../command/guards';
 import { ParsedEnvVariables } from '../../configuration/env/types';
 import { RefreshingAuthProvider } from '@twurple/auth';
+import { setTimeoutPromise } from '../../utils/promise';
 
 /**
  * A closure that returns a ChatClient onMessageHandler to call the commands and provide access control
@@ -17,23 +18,48 @@ import { RefreshingAuthProvider } from '@twurple/auth';
  */
 export const onMessageHandlerClosure = (authProvider: RefreshingAuthProvider, chatClient: ChatClient, env: ParsedEnvVariables, source: RequestSource): MessageHandler => {
     return async (channel, user, message, msg): Promise<void> => {
-        const [command, args = GLOBAL.EMPTY_MESSAGE] = getCommand(message);
-        // Ignore messages that are not commands
-        if (command == null) {
-            return;
-        }
+        const commandList = getCommandList(message);
         try {
-            const commandHandler = CommandHandlers[command] as CommandHandlerType;
-            // If rewards mode enabled and the input is a command and the user is not streamer or mod or vip, only allow safe commands
-            if (commandHandler == null) {
-                return;
-            }
+            // Start all tasks asynchronously
+            const promiseList = commandList.map(async ([command, args, delayNs]) => {
+                // Ignore messages that are not commands
+                if (command == null) {
+                    return;
+                }
 
-            // If no user info was provided, this is is Channel Points/Rewards mode, so there's no block
-            const twitch: TwitchParams = { channel, chatClient, authProvider, user, broadcasterUser: env.TARGET_CHANNEL, userRoles: msg?.userInfo ?? CONFIG.DEFAULT_USER_ROLES };
-            // Checks if the user has enough permissions
-            checkCommandAccess(command, twitch, source, env);
-            await commandHandler(args, { targetMIDIChannel: env.TARGET_MIDI_CHANNEL, targetMIDIName: env.TARGET_MIDI_NAME, isRewardsMode: env.REWARDS_MODE }, twitch);
+                // Delay before instruction
+                await setTimeoutPromise(delayNs);
+
+                const commandHandler = CommandHandlers[command] as CommandHandlerType;
+                // If rewards mode enabled and the input is a command and the user is not streamer or mod or vip, only allow safe commands
+                if (commandHandler == null) {
+                    return;
+                }
+
+                // If no user info was provided, this is is Channel Points/Rewards mode, so there's no block
+                const twitch: TwitchParams = {
+                    channel,
+                    chatClient,
+                    authProvider,
+                    user,
+                    broadcasterUser: env.TARGET_CHANNEL,
+                    userRoles: msg?.userInfo ?? CONFIG.DEFAULT_USER_ROLES
+                };
+                // Checks if the user has enough permissions
+                checkCommandAccess(command, twitch, source, env);
+                // Call command
+                return commandHandler(
+                    args,
+                    {
+                        targetMIDIChannel: env.TARGET_MIDI_CHANNEL,
+                        targetMIDIName: env.TARGET_MIDI_NAME,
+                        isRewardsMode: env.REWARDS_MODE
+                    },
+                    twitch
+                );
+            });
+            // Collect all promises and their errors
+            await Promise.all(promiseList);
         } catch (error) {
             // Skip error notification if SEND_UNAUTHORIZED_MESSAGE is false
             if (!(error instanceof Error) || error.message !== ERROR_MSG.BAD_PERMISSIONS() || env.SEND_UNAUTHORIZED_MESSAGE) {